**Automatic testing of applications on Flash.**
Step 1. Create a file structure for the tests.

Step 2. Create a fixture with a mac for Directorydao.

Step 3. Writing a class with tests for DirectorService.

Step 4. Create a fixture with a mock for GenreDAO.

Step 5. Writing a class with tests for GenreService.

Step 6. Create a mock fixture for MovieDAO.

Step 7. Writing a class with tests for MovieService.